+++ src/events/page/sessions/list/EventSessions.tsx
@@ -15,7 +15,6 @@ import {
     TextField,
     Typography,
     Checkbox,
-    Alert,
 } from '@mui/material'
 import * as React from 'react'
 import { useMemo, useState } from 'react'
@@ -34,7 +33,7 @@ import { useSearchParams } from '../../../../hooks/useSearchParams'
 import { GenerateSessionsVideoDialog } from '../components/GenerateSessionsVideoDialog'
 import { exportSessionsAction, SessionsExportType } from './actions/exportSessionsActions'
 import { TeasingPostSocials } from '../../../actions/sessions/generation/generateSessionTeasingContent'
-import { useMoveAllImagesToOpenPlannerStorage } from '../../../actions/speakers/useMoveAllImagesToOpenPlannerStorage'
+import { MoveImagesAlert } from '../components/MoveImagesAlert'
 
 export type EventSessionsProps = {
     event: Event
@@ -60,8 +59,6 @@ export const EventSessions = ({ event }: EventSessionsProps) => {
     }
 
     const sessionsData = useMemo(() => sessions.data || [], [sessions.data])
-
-    const moveAllImagesToOpenPlannerStorageState = useMoveAllImagesToOpenPlannerStorage(event, sessionsData)
 
     const displayedSessions = useMemo(() => {
         return filterSessions(sessionsData, {
@@ -103,40 +100,8 @@ export const EventSessions = ({ event }: EventSessionsProps) => {
 
     return (
         <Container maxWidth="lg" sx={{ mt: 4, mb: 4 }}>
-            {moveAllImagesToOpenPlannerStorageState.shouldUpdateImageStorage && (
-                <Alert variant="filled" severity="info" sx={{ marginBottom: 2 }}>
-                    {moveAllImagesToOpenPlannerStorageState.total} session/speaker images are stored in OpenPlanner but
-                    not in the storage bucket. This may cause issues due to the image not being accessible in some code
-                    (ShortVid.io) or being moved later on making it unavailable.
-                    <br />
-                    You can migrate the images to OpenPlanner at once now:
-                    <Button
-                        variant="contained"
-                        color="secondary"
-                        disabled={moveAllImagesToOpenPlannerStorageState.isLoading}
-                        onClick={() => {
-                            moveAllImagesToOpenPlannerStorageState.moveAllImagesToOpenPlannerStorage()
-                        }}>
-                        Migrate images{' '}
-                        {moveAllImagesToOpenPlannerStorageState.isLoading
-                            ? `${moveAllImagesToOpenPlannerStorageState.progress}/${moveAllImagesToOpenPlannerStorageState.total}`
-                            : ''}
-                    </Button>
-                    {moveAllImagesToOpenPlannerStorageState.error.length > 0 && (
-                        <>
-                            <Typography>Error(s):</Typography>
-                            <ul>
-                                {moveAllImagesToOpenPlannerStorageState.error.map((error) => (
-                                    <li key={error}>{error}</li>
-                                ))}
-                            </ul>
-                            <Typography>
-                                Common errors could be linked to CORS issues, please check the console for more details.
-                            </Typography>
-                        </>
-                    )}
-                </Alert>
-            )}
+            <MoveImagesAlert event={event} sessionsData={sessionsData} />
+
             <Box display="flex" justifyContent="space-between" alignItems="center" flexWrap="wrap" marginBottom={1}>
                 <Typography>
                     {isFiltered
